pyproject.toml
[project]
name = "payment-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Kredo15",email = "kredo1511@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiogram-i18n (>=1.4,<2.0)",
    "fluent-compiler (>=1.1,<2.0)",
    "fluent-runtime (>=0.4.0,<0.5.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "aiocryptopay (>=0.4.8,<0.5.0)",
    "redis (>=6.4.0,<7.0.0)",
    "yookassa (>=3.7.1,<4.0.0)",
    "black (>=25.9.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
